requirements.txt
cbox
h5py
scipy
arrow
Keras
numpy
Pillow
pytest
imageio
moviepy
scikit-learn
tensorflow
tensorflow-tensorboard
matplotlib
requests
urllib3
ipython
ipython-genutils
soocii-services-lib
google-cloud
google-auth-oauthlib
